Move module `heuristics` in subfolder `src/bed_bpp_env`
